Security Improvements; Fail2Ban Support; Don't Report Username
Don't distinguish between a bad email address and a bad password
in the error response to the user.  We don't need to make the
hackers job any easier.

Log failed Auth attempts to the error log.  This allows for the
use of Fail2Ban to block ip addresses associated with multiple
failed attempts.  Otherwise, hackers are free to brute force the
email/password field without any limits.